chore: master 覆盖为 codeBrave-World-node2 快照
按用户确认，将默认分支整树替换为功能分支快照（含 tslua2→Brave-World 改名）。
master 独有的 a55960d（技能特效/开发工具链/网关重构）内容被移除，但其提交仍保留在历史中，
可用 backup/master-before-overwrite 或 git revert 本提交恢复。

+++ servercsharp/src/GameServer.Common/Buffs/BuffContainer.cs
@@ -96,8 +96,8 @@ public class BuffContainer
 
     // ---- 属性修正 ----
 
-    /// <summary>获取某属性的修正值总和</summary>
-    public int GetAttrModifier(string attrName)
+    /// <summary>获取某属性的修正值总和（返回 double 避免百分比修正被截断）</summary>
+    public double GetAttrModifier(string attrName)
     {
         double total = 0;
         foreach (var buff in _buffs)
@@ -113,7 +113,7 @@ public class BuffContainer
                     total += mod.Value * buff.Stacks;
             }
         }
-        return (int)total;
+        return total;
     }
 
     // ---- 护盾 ----
